flask-tools-site-clean/requirements.txt
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==45.0.2
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
docx2pdf==0.1.8
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.2
fire==0.7.0
Flask==3.1.0
Flask-Caching==2.3.1
fonttools==4.58.0
fpdf==1.7.2
gunicorn
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.4
moviepy==2.2.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.6
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.5
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1

platformdirs==4.4.0
proglog==0.1.12
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
PyMuPDF==1.25.5
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
pyzmq==27.0.2
qrcode==8.2
referencing==0.36.2
reportlab==3.6.13
requests==2.32.5
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
termcolor==3.1.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
xlsx2pdf==1.0.4
XlsxWriter==3.2.3
yarg==0.1.9
yt-dlp==2025.5.22
